// src/app/generated/angular-client/model/functionMasterDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductMasterDto } from './productMasterDto';


/**
 * Function Master DTO
 */
export interface FunctionMasterDto { 
    id?: string;
    functionName?: string;
    createdBy?: string;
    createdAt?: string;
    productMasterDtos?: Array<ProductMasterDto>;
}